add upgrade test
