+++ lucene/backward-codecs/src/test/org/apache/lucene/backward_index/TestBackwardsCompatibility.java
@@ -181,38 +181,7 @@ public class TestBackwardsCompatibility extends LuceneTestCase {
     }
   }
 
-  static final String[] oldNames = {
-    "9.0.0-cfs", // Force on separate lines
-    "9.0.0-nocfs",
-    "9.1.0-cfs",
-    "9.1.0-nocfs",
-    "9.2.0-cfs",
-    "9.2.0-nocfs",
-    "9.3.0-cfs",
-    "9.3.0-nocfs",
-    "9.4.0-cfs",
-    "9.4.0-nocfs",
-    "9.4.1-cfs",
-    "9.4.1-nocfs",
-    "9.4.2-cfs",
-    "9.4.2-nocfs",
-    "9.5.0-cfs",
-    "9.5.0-nocfs",
-    "9.6.0-cfs",
-    "9.6.0-nocfs",
-    "9.7.0-cfs",
-    "9.7.0-nocfs",
-    "9.8.0-cfs",
-    "9.8.0-nocfs",
-    "9.9.0-cfs",
-    "9.9.0-nocfs",
-    "9.9.1-cfs",
-    "9.9.1-nocfs"
-  };
 
-  public static String[] getOldNames() {
-    return oldNames;
-  }
 
   static final String[] unsupportedNames = {
     "1.9.0-cfs",
@@ -478,170 +447,10 @@ public class TestBackwardsCompatibility extends LuceneTestCase {
     binarySupportedNames = list.toArray(new String[0]);
   }
 
-  // TODO: on 6.0.0 release, gen the single segment indices and add here:
-  static final String[] oldSingleSegmentNames = {};
-
-  public static String[] getOldSingleSegmentNames() {
-    return oldSingleSegmentNames;
-  }
-
   static Map<String, Directory> oldIndexDirs;
 
-  /** Randomizes the use of some of hte constructor variations */
-  private static IndexUpgrader newIndexUpgrader(Directory dir) {
-    final boolean streamType = random().nextBoolean();
-    final int choice = TestUtil.nextInt(random(), 0, 2);
-    switch (choice) {
-      case 0:
-        return new IndexUpgrader(dir);
-      case 1:
-        return new IndexUpgrader(dir, streamType ? null : InfoStream.NO_OUTPUT, false);
-      case 2:
-        return new IndexUpgrader(dir, newIndexWriterConfig(null), false);
-      default:
-        fail("case statement didn't get updated when random bounds changed");
-    }
-    return null; // never get here
-  }
 
-  @BeforeClass
-  public static void beforeClass() throws Exception {
-    List<String> names = new ArrayList<>(oldNames.length + oldSingleSegmentNames.length);
-    names.addAll(Arrays.asList(oldNames));
-    names.addAll(Arrays.asList(oldSingleSegmentNames));
-    oldIndexDirs = new HashMap<>();
-    for (String name : names) {
-      Path dir = createTempDir(name);
-      InputStream resource =
-          TestBackwardsCompatibility.class.getResourceAsStream("index." + name + ".zip");
-      assertNotNull("Index name " + name + " not found", resource);
-      TestUtil.unzip(resource, dir);
-      oldIndexDirs.put(name, newFSDirectory(dir));
-    }
-  }
 
-  @AfterClass
-  public static void afterClass() throws Exception {
-    for (Directory d : oldIndexDirs.values()) {
-      d.close();
-    }
-    oldIndexDirs = null;
-  }
-
-  public void testAllVersionHaveCfsAndNocfs() {
-    // ensure all tested versions with cfs also have nocfs
-    String[] files = new String[oldNames.length];
-    System.arraycopy(oldNames, 0, files, 0, oldNames.length);
-    Arrays.sort(files);
-    String prevFile = "";
-    for (String file : files) {
-      if (prevFile.endsWith("-cfs")) {
-        String prefix = prevFile.replace("-cfs", "");
-        assertEquals("Missing -nocfs for backcompat index " + prefix, prefix + "-nocfs", file);
-      }
-      prevFile = file;
-    }
-  }
-
-  public void testAllVersionsTested() throws Exception {
-    Pattern constantPattern = Pattern.compile("LUCENE_(\\d+)_(\\d+)_(\\d+)(_ALPHA|_BETA)?");
-    // find the unique versions according to Version.java
-    List<String> expectedVersions = new ArrayList<>();
-    for (java.lang.reflect.Field field : Version.class.getDeclaredFields()) {
-      if (Modifier.isStatic(field.getModifiers()) && field.getType() == Version.class) {
-        Version v = (Version) field.get(Version.class);
-        if (v.equals(Version.LATEST)) {
-          continue;
-        }
-
-        Matcher constant = constantPattern.matcher(field.getName());
-        if (constant.matches() == false) {
-          continue;
-        }
-
-        expectedVersions.add(v + "-cfs");
-      }
-    }
-
-    // BEGIN TRUNK ONLY BLOCK
-    // on trunk, the last release of the prev major release is also untested
-    Version lastPrevMajorVersion = null;
-    for (java.lang.reflect.Field field : Version.class.getDeclaredFields()) {
-      if (Modifier.isStatic(field.getModifiers()) && field.getType() == Version.class) {
-        Version v = (Version) field.get(Version.class);
-        Matcher constant = constantPattern.matcher(field.getName());
-        if (constant.matches() == false) continue;
-        if (v.major == Version.LATEST.major - 1
-            && (lastPrevMajorVersion == null || v.onOrAfter(lastPrevMajorVersion))) {
-          lastPrevMajorVersion = v;
-        }
-      }
-    }
-    assertNotNull(lastPrevMajorVersion);
-    expectedVersions.remove(lastPrevMajorVersion + "-cfs");
-    // END TRUNK ONLY BLOCK
-
-    Collections.sort(expectedVersions);
-
-    // find what versions we are testing
-    List<String> testedVersions = new ArrayList<>();
-    for (String testedVersion : oldNames) {
-      if (testedVersion.endsWith("-cfs") == false) {
-        continue;
-      }
-      testedVersions.add(testedVersion);
-    }
-    Collections.sort(testedVersions);
-
-    int i = 0;
-    int j = 0;
-    List<String> missingFiles = new ArrayList<>();
-    List<String> extraFiles = new ArrayList<>();
-    while (i < expectedVersions.size() && j < testedVersions.size()) {
-      String expectedVersion = expectedVersions.get(i);
-      String testedVersion = testedVersions.get(j);
-      int compare = expectedVersion.compareTo(testedVersion);
-      if (compare == 0) { // equal, we can move on
-        ++i;
-        ++j;
-      } else if (compare < 0) { // didn't find test for version constant
-        missingFiles.add(expectedVersion);
-        ++i;
-      } else { // extra test file
-        extraFiles.add(testedVersion);
-        ++j;
-      }
-    }
-    while (i < expectedVersions.size()) {
-      missingFiles.add(expectedVersions.get(i));
-      ++i;
-    }
-    while (j < testedVersions.size()) {
-      missingFiles.add(testedVersions.get(j));
-      ++j;
-    }
-
-    // we could be missing up to 1 file, which may be due to a release that is in progress
-    if (missingFiles.size() <= 1 && extraFiles.isEmpty()) {
-      // success
-      return;
-    }
-
-    StringBuilder msg = new StringBuilder();
-    if (missingFiles.size() > 1) {
-      msg.append("Missing backcompat test files:\n");
-      for (String missingFile : missingFiles) {
-        msg.append("  ").append(missingFile).append("\n");
-      }
-    }
-    if (extraFiles.isEmpty() == false) {
-      msg.append("Extra backcompat test files:\n");
-      for (String extraFile : extraFiles) {
-        msg.append("  ").append(extraFile).append("\n");
-      }
-    }
-    fail(msg.toString());
-  }
 
   /**
    * This test checks that *only* IndexFormatTooOldExceptions are thrown when you open and operate
@@ -775,117 +584,7 @@ public class TestBackwardsCompatibility extends LuceneTestCase {
     }
   }
 
-  public void testFullyMergeOldIndex() throws Exception {
-    for (String name : oldNames) {
-      if (VERBOSE) {
-        System.out.println("\nTEST: index=" + name);
-      }
-      Directory dir = newDirectory(oldIndexDirs.get(name));
 
-      final SegmentInfos oldSegInfos = SegmentInfos.readLatestCommit(dir);
-
-      IndexWriter w = new IndexWriter(dir, new IndexWriterConfig(new MockAnalyzer(random())));
-      w.forceMerge(1);
-      w.close();
-
-      final SegmentInfos segInfos = SegmentInfos.readLatestCommit(dir);
-      assertEquals(
-          oldSegInfos.getIndexCreatedVersionMajor(), segInfos.getIndexCreatedVersionMajor());
-      assertEquals(Version.LATEST, segInfos.asList().get(0).info.getVersion());
-      assertEquals(
-          oldSegInfos.asList().get(0).info.getMinVersion(),
-          segInfos.asList().get(0).info.getMinVersion());
-
-      dir.close();
-    }
-  }
-
-  public void testAddOldIndexes() throws IOException {
-    for (String name : oldNames) {
-      if (VERBOSE) {
-        System.out.println("\nTEST: old index " + name);
-      }
-      Directory oldDir = oldIndexDirs.get(name);
-      SegmentInfos infos = SegmentInfos.readLatestCommit(oldDir);
-
-      Directory targetDir = newDirectory();
-      if (infos.getCommitLuceneVersion().major != Version.LATEST.major) {
-        // both indexes are not compatible
-        Directory targetDir2 = newDirectory();
-        IndexWriter w =
-            new IndexWriter(targetDir2, newIndexWriterConfig(new MockAnalyzer(random())));
-        IllegalArgumentException e =
-            expectThrows(IllegalArgumentException.class, () -> w.addIndexes(oldDir));
-        assertTrue(
-            e.getMessage(),
-            e.getMessage()
-                .startsWith(
-                    "Cannot use addIndexes(Directory) with indexes that have been created by a different Lucene version."));
-        w.close();
-        targetDir2.close();
-
-        // for the next test, we simulate writing to an index that was created on the same major
-        // version
-        new SegmentInfos(infos.getIndexCreatedVersionMajor()).commit(targetDir);
-      }
-
-      IndexWriter w = new IndexWriter(targetDir, newIndexWriterConfig(new MockAnalyzer(random())));
-      w.addIndexes(oldDir);
-      w.close();
-
-      SegmentInfos si = SegmentInfos.readLatestCommit(targetDir);
-      assertNull(
-          "none of the segments should have been upgraded",
-          si.asList().stream()
-              .filter( // depending on the MergePolicy we might see these segments merged away
-                  sci ->
-                      sci.getId() != null
-                          && sci.info.getVersion().onOrAfter(Version.fromBits(8, 6, 0)) == false)
-              .findAny()
-              .orElse(null));
-      if (VERBOSE) {
-        System.out.println("\nTEST: done adding indices; now close");
-      }
-
-      targetDir.close();
-    }
-  }
-
-  public void testAddOldIndexesReader() throws IOException {
-    for (String name : oldNames) {
-      Directory oldDir = oldIndexDirs.get(name);
-      SegmentInfos infos = SegmentInfos.readLatestCommit(oldDir);
-      DirectoryReader reader = DirectoryReader.open(oldDir);
-
-      Directory targetDir = newDirectory();
-      if (infos.getCommitLuceneVersion().major != Version.LATEST.major) {
-        Directory targetDir2 = newDirectory();
-        IndexWriter w =
-            new IndexWriter(targetDir2, newIndexWriterConfig(new MockAnalyzer(random())));
-        IllegalArgumentException e =
-            expectThrows(
-                IllegalArgumentException.class, () -> TestUtil.addIndexesSlowly(w, reader));
-        assertEquals(
-            e.getMessage(),
-            "Cannot merge a segment that has been created with major version 9 into this index which has been created by major version 10");
-        w.close();
-        targetDir2.close();
-
-        // for the next test, we simulate writing to an index that was created on the same major
-        // version
-        new SegmentInfos(infos.getIndexCreatedVersionMajor()).commit(targetDir);
-      }
-      IndexWriter w = new IndexWriter(targetDir, newIndexWriterConfig(new MockAnalyzer(random())));
-      TestUtil.addIndexesSlowly(w, reader);
-      w.close();
-      reader.close();
-      SegmentInfos si = SegmentInfos.readLatestCommit(targetDir);
-      assertNull(
-          "all SCIs should have an id now",
-          si.asList().stream().filter(sci -> sci.getId() == null).findAny().orElse(null));
-      targetDir.close();
-    }
-  }
 
 
 
@@ -908,224 +607,19 @@ public class TestBackwardsCompatibility extends LuceneTestCase {
     r.close();
   }
 
-
-  private int checkAllSegmentsUpgraded(Directory dir, int indexCreatedVersion) throws IOException {
-    final SegmentInfos infos = SegmentInfos.readLatestCommit(dir);
-    if (VERBOSE) {
-      System.out.println("checkAllSegmentsUpgraded: " + infos);
-    }
-    for (SegmentCommitInfo si : infos) {
-      assertEquals(Version.LATEST, si.info.getVersion());
-      assertNotNull(si.getId());
-    }
-    assertEquals(Version.LATEST, infos.getCommitLuceneVersion());
-    assertEquals(indexCreatedVersion, infos.getIndexCreatedVersionMajor());
-    return infos.size();
-  }
-
-  private int getNumberOfSegments(Directory dir) throws IOException {
-    final SegmentInfos infos = SegmentInfos.readLatestCommit(dir);
-    return infos.size();
-  }
-
-  public void testUpgradeOldIndex() throws Exception {
-    List<String> names = new ArrayList<>(oldNames.length + oldSingleSegmentNames.length);
-    names.addAll(Arrays.asList(oldNames));
-    names.addAll(Arrays.asList(oldSingleSegmentNames));
-    for (String name : names) {
-      if (VERBOSE) {
-        System.out.println("testUpgradeOldIndex: index=" + name);
-      }
-      Directory dir = newDirectory(oldIndexDirs.get(name));
-      int indexCreatedVersion = SegmentInfos.readLatestCommit(dir).getIndexCreatedVersionMajor();
-
-      newIndexUpgrader(dir).upgrade();
-
-      checkAllSegmentsUpgraded(dir, indexCreatedVersion);
-
-      dir.close();
-    }
-  }
-
-  public void testIndexUpgraderCommandLineArgs() throws Exception {
-
-    PrintStream savedSystemOut = System.out;
-    System.setOut(new PrintStream(new ByteArrayOutputStream(), false, UTF_8));
-    try {
-      for (Map.Entry<String, Directory> entry : oldIndexDirs.entrySet()) {
-        String name = entry.getKey();
-        Directory origDir = entry.getValue();
-        int indexCreatedVersion =
-            SegmentInfos.readLatestCommit(origDir).getIndexCreatedVersionMajor();
-        Path dir = createTempDir(name);
-        try (FSDirectory fsDir = FSDirectory.open(dir)) {
-          // beware that ExtraFS might add extraXXX files
-          Set<String> extraFiles = Set.of(fsDir.listAll());
-          for (String file : origDir.listAll()) {
-            if (extraFiles.contains(file) == false) {
-              fsDir.copyFrom(origDir, file, file, IOContext.DEFAULT);
-            }
-          }
-        }
-
-        String path = dir.toAbsolutePath().toString();
-
-        List<String> args = new ArrayList<>();
-        if (random().nextBoolean()) {
-          args.add("-verbose");
-        }
-        if (random().nextBoolean()) {
-          args.add("-delete-prior-commits");
-        }
-        if (random().nextBoolean()) {
-          // TODO: need to better randomize this, but ...
-          //  - LuceneTestCase.FS_DIRECTORIES is private
-          //  - newFSDirectory returns BaseDirectoryWrapper
-          //  - BaseDirectoryWrapper doesn't expose delegate
-          Class<? extends FSDirectory> dirImpl = NIOFSDirectory.class;
-
-          args.add("-dir-impl");
-          args.add(dirImpl.getName());
-        }
-        args.add(path);
-
-        IndexUpgrader.main(args.toArray(new String[0]));
-
-        Directory upgradedDir = newFSDirectory(dir);
-        try (upgradedDir) {
-          checkAllSegmentsUpgraded(upgradedDir, indexCreatedVersion);
-        }
-      }
-    } finally {
-      System.setOut(savedSystemOut);
-    }
-  }
-
-  public void testUpgradeOldSingleSegmentIndexWithAdditions() throws Exception {
-    for (String name : oldSingleSegmentNames) {
-      if (VERBOSE) {
-        System.out.println("testUpgradeOldSingleSegmentIndexWithAdditions: index=" + name);
-      }
-      Directory dir = newDirectory(oldIndexDirs.get(name));
-      assertEquals("Original index must be single segment", 1, getNumberOfSegments(dir));
-      int indexCreatedVersion = SegmentInfos.readLatestCommit(dir).getIndexCreatedVersionMajor();
-
-      // create a bunch of dummy segments
-      int id = 40;
-      Directory ramDir = new ByteBuffersDirectory();
-      for (int i = 0; i < 3; i++) {
-        // only use Log- or TieredMergePolicy, to make document addition predictable and not
-        // suddenly merge:
-        MergePolicy mp = random().nextBoolean() ? newLogMergePolicy() : newTieredMergePolicy();
-        IndexWriterConfig iwc =
-            new IndexWriterConfig(new MockAnalyzer(random())).setMergePolicy(mp);
-        IndexWriter w = new IndexWriter(ramDir, iwc);
-        // add few more docs:
-        for (int j = 0; j < RANDOM_MULTIPLIER * random().nextInt(30); j++) {
-          TestBasicBackwardsCompatibility.addDoc(w, id++);
-        }
-        try {
-          w.commit();
-        } finally {
-          w.close();
-        }
-      }
-
-      // add dummy segments (which are all in current
-      // version) to single segment index
-      MergePolicy mp = random().nextBoolean() ? newLogMergePolicy() : newTieredMergePolicy();
-      IndexWriterConfig iwc = new IndexWriterConfig(null).setMergePolicy(mp);
-      IndexWriter w = new IndexWriter(dir, iwc);
-      w.addIndexes(ramDir);
-      try (w) {
-        w.commit();
-      }
-
-      // determine count of segments in modified index
-      final int origSegCount = getNumberOfSegments(dir);
-
-      // ensure there is only one commit
-      assertEquals(1, DirectoryReader.listCommits(dir).size());
-      newIndexUpgrader(dir).upgrade();
-
-      final int segCount = checkAllSegmentsUpgraded(dir, indexCreatedVersion);
-      assertEquals(
-          "Index must still contain the same number of segments, as only one segment was upgraded and nothing else merged",
-          origSegCount,
-          segCount);
-
-      dir.close();
-    }
-  }
-
   public static final String emptyIndex = "empty.9.0.0.zip";
 
   public void testUpgradeEmptyOldIndex() throws Exception {
     Path oldIndexDir = createTempDir("emptyIndex");
     TestUtil.unzip(getDataInputStream(emptyIndex), oldIndexDir);
     Directory dir = newFSDirectory(oldIndexDir);
+    TestIndexUpgradeBackwardsCompatibility.newIndexUpgrader(dir).upgrade();
 
-    newIndexUpgrader(dir).upgrade();
-
-    checkAllSegmentsUpgraded(dir, 9);
+    TestIndexUpgradeBackwardsCompatibility.checkAllSegmentsUpgraded(dir, 9);
 
     dir.close();
   }
 
-
-  // LUCENE-5907
-  public void testUpgradeWithNRTReader() throws Exception {
-    for (String name : oldNames) {
-      Directory dir = newDirectory(oldIndexDirs.get(name));
-
-      IndexWriter writer =
-          new IndexWriter(
-              dir, newIndexWriterConfig(new MockAnalyzer(random())).setOpenMode(OpenMode.APPEND));
-      writer.addDocument(new Document());
-      DirectoryReader r = DirectoryReader.open(writer);
-      writer.commit();
-      r.close();
-      writer.forceMerge(1);
-      writer.commit();
-      writer.rollback();
-      SegmentInfos.readLatestCommit(dir);
-      dir.close();
-    }
-  }
-
-  // LUCENE-5907
-  public void testUpgradeThenMultipleCommits() throws Exception {
-    for (String name : oldNames) {
-      Directory dir = newDirectory(oldIndexDirs.get(name));
-
-      IndexWriter writer =
-          new IndexWriter(
-              dir, newIndexWriterConfig(new MockAnalyzer(random())).setOpenMode(OpenMode.APPEND));
-      writer.addDocument(new Document());
-      writer.commit();
-      writer.addDocument(new Document());
-      writer.commit();
-      writer.close();
-      dir.close();
-    }
-  }
-
-  public void testFailOpenOldIndex() throws IOException {
-    for (String name : oldNames) {
-      Directory directory = oldIndexDirs.get(name);
-      IndexCommit commit = DirectoryReader.listCommits(directory).get(0);
-      IndexFormatTooOldException ex =
-          expectThrows(
-              IndexFormatTooOldException.class,
-              () -> StandardDirectoryReader.open(commit, Version.LATEST.major, null));
-      assertTrue(
-          ex.getMessage()
-              .contains(
-                  "only supports reading from version " + Version.LATEST.major + " upwards."));
-      // now open with allowed min version
-      StandardDirectoryReader.open(commit, Version.MIN_SUPPORTED_MAJOR, null).close();
-    }
-  }
 
   // #12895: test on a carefully crafted 9.8.0 index (from a small contiguous subset
   // of wikibigall unique terms) that shows the read-time exception of
@@ -1172,39 +666,4 @@ public class TestBackwardsCompatibility extends LuceneTestCase {
     }
   }
 
-  public void testOpenModeAndCreatedVersion() throws IOException {
-    for (String name : oldNames) {
-      Directory dir = newDirectory(oldIndexDirs.get(name));
-      int majorVersion = SegmentInfos.readLatestCommit(dir).getIndexCreatedVersionMajor();
-      if (majorVersion != Version.MIN_SUPPORTED_MAJOR && majorVersion != Version.LATEST.major) {
-        fail(
-            "expected one of: ["
-                + Version.MIN_SUPPORTED_MAJOR
-                + ", "
-                + Version.LATEST.major
-                + "] but got: "
-                + majorVersion);
-      }
-      for (OpenMode openMode : OpenMode.values()) {
-        Directory tmpDir = newDirectory(dir);
-        IndexWriter w = new IndexWriter(tmpDir, newIndexWriterConfig().setOpenMode(openMode));
-        w.commit();
-        w.close();
-        switch (openMode) {
-          case CREATE:
-            assertEquals(
-                Version.LATEST.major,
-                SegmentInfos.readLatestCommit(tmpDir).getIndexCreatedVersionMajor());
-            break;
-          case APPEND:
-          case CREATE_OR_APPEND:
-          default:
-            assertEquals(
-                majorVersion, SegmentInfos.readLatestCommit(tmpDir).getIndexCreatedVersionMajor());
-        }
-        tmpDir.close();
-      }
-      dir.close();
-    }
-  }
 }
